Reorganized the project filesystem of the get location weather api route
