+++ src/api/weatherApi.ts
@@ -1,25 +1,5 @@
-interface WeatherApiResponse {
-  cod: number;
-  main: {
-    temp: number;
-    humidity: number;
-  };
-  wind: {
-    speed: number;
-  };
-  rain: { [key: string]: number };
-  weather: {
-    description: string;
-  }[];
-}
 
-interface ExtractedWeatherInfo {
-  temperature: number;
-  humidity: number;
-  windSpeed: number;
-  precipitation: number | null;
-  weatherDescription: string | null;
-}
+import { extractWeatherInfo } from "../utilities/weatherUtils";
 
 export async function getLocationWeather() {
   const res = await fetch(
@@ -42,31 +22,4 @@ export async function getLocationWeather() {
   const weatherInfo = extractWeatherInfo(data);
 
   return weatherInfo;
-}
-
-function extractWeatherInfo(apiResponse: WeatherApiResponse) {
-  if (!apiResponse || apiResponse.cod !== 200) {
-    throw new Error("Invalid API Response");
-  }
-
-  const {
-    main: { temp, humidity },
-    wind: { speed },
-    weather,
-    rain
-  } = apiResponse;
-
-  const mainWeather = weather && weather.length > 0 ? weather[0] : null;
-  const weatherDescription = mainWeather ? mainWeather.description : null;
-  const precipitation = rain ? rain["1h"] : null;
-
-  const extractedInfo: ExtractedWeatherInfo = {
-    temperature: temp,
-    humidity,
-    windSpeed: speed,
-    precipitation,
-    weatherDescription
-  };
-
-  return extractedInfo;
 }
